syscfg: Do not rely on generated sysconfig
Instead of blindly relying on whatever CCS/Sysconfig generates, lets
commit it to the repository. This ensures that every build will be the
same.

Changes will still require to be managed by a human in the end, but this
should be done by one of the developers, instead of everybody for
themselves.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>