patches: Move preinclude to the source
Instead of patching the workspace, move preinclude into the projectspec
and SDK sources instead. This allows to get everything when importing a
project, reducing the need to patch the workspace a bit.

Note, that this commit breaks support for 'other' platform, by no longer
offering the preinclude define hack. Also note that support is not
broken, just that the GPIO's need to be correctly set in the
IDE/znp.syscfg.

This is intended and will be resolved in a next patch in this series.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>