syscfg: Do not rely on generated sysconfig
Instead of blindly relying on whatever CCS/Sysconfig generates, lets
commit it to the repository. This ensures that every build will be the
same. This also means eclipse (GUI) is no longer needed if commands are
executed shell.

Changes will still require to be managed by a human in the end, but this
should be done by one of the developers, instead of everybody for
themselves.

Ideally, these files would be generated at runtime using sysconfig
instead, but at this time that was not possible/figured out yet. See [0]
for more information.

[0]: https://e2e.ti.com/support/wireless-connectivity/zigbee-thread-group/zigbee-and-thread/f/zigbee-thread-forum/1397346/cc2652p-using-sysconfig-from-the-cli-to-generate-znp-sysfg

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>