fix: Set txPower to 9 to match hardcoded config
There's no point in setting the txPower to 20 in our config, when in fact
the txPower is set to 9 via a hardcoded define in this very same patch.

Instead, ensure that the generated files have the proper txPower defined
so that the `ti_zstack_config.h` [0] can set it properly internally.

[0]: source/ti/zstack/.meta/templates/ti_zstack_config.h.xdt:463

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>